requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
python-docx==1.0.1
Pillow==10.0.0
pytesseract==0.3.10
nltk==3.8.1
scikit-learn==1.3.0
spacy==3.7.2
gensim==4.3.2
gunicorn==21.2.0
pandas==2.1.1
numpy==1.25.2
matplotlib==3.8.0
joblib==1.3.2
setuptools>=65.0.0
Werkzeug>=2.3.7
